requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
black==24.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.7
django-admin-adminlte==1.0.5
django-adminlte3==0.1.6
django-allauth==0.60.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwcrypto==1.5.0
kombu==5.3.5
msgpack==1.0.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
stripe==7.13.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
zope.interface==6.1
